Fix dtype implementation and testing in denoise_bilateral

+++ skimage/restoration/tests/test_denoise.py
@@ -195,7 +195,7 @@ def test_denoise_bilateral_types(dtype):
     img = checkerboard_gray.copy()[:50, :50]
     # add some random noise
     img += 0.5 * img.std() * np.random.rand(*img.shape)
-    img = np.clip(img, 0, 1)
+    img = np.clip(img, 0, 1).astype(dtype)
 
     # check that we can process multiple float types
     out = restoration.denoise_bilateral(img, sigma_color=0.1,
